admin/requirements.txt
python-dotenv

Flask
boto3
pillow
